requirements-discord-bot.txt
py-cord>=2.4.0
python-dotenv>=1.0.0
aiohttp>=3.8.0